+++ HISTORY.md
@@ -5,6 +5,7 @@
 
 - Fixed the `babble` module always sending messages via it's own internal
   messagebus instead of the agents send/receive function.
+- WebSocketTransport accepts any id.
 - Improved error handling for sending messages.